Started implementing a new uid method

+++ include/ccNoise/ccNoise.h
@@ -20,13 +20,15 @@
 
 #pragma once
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
 // Create an unique as possible ID for a coordinate
-unsigned int ccnCoordinateUid(unsigned int dimensions, int *coordinate);
+uint64_t ccnCoordinateUid(unsigned int dimensions, int *coordinate);
 
 // Creates n-dimensional perlin noise
 // persistence 0 - 0.5
